P0-2: store money as integer cents (BIGINT) to eliminate float rounding drift
- 001_initial.sql: amount_without_vat/vat_amount/amount_with_vat,
  unit_price/total_price/vat_amount REAL -> BIGINT (cents); vat_rate,
  tariff, quantity stay REAL (percent/quantity, not money). No migration:
  project not yet in use.
- domain/models.go: money fields *float64 -> *int64 (cents).
- worker.go: toCents(float64) converts AI-processor euro floats to cents
  at INSERT boundary (invoice totals + items; total_price = gross).
- httpapi/review.go: parseCents(eur string) *int64 parses form euro
  values to cents (money fields); parseFloat kept for quantity/vat_rate.
- export: centsToFloat(*int64) helper converts cents -> euros for the
  export payload (templates unchanged, fed floats); BuildPayload uses it.
- webui: centsToFloat template func; money fields in invoices/company_details/
  home/review templates switched from derefFloat to centsToFloat; quantity
  and vat_rate keep derefFloat.
- tests: cents/centsPtr helpers; DB total_price assertion in cents (12100);
  service/httpapi VatAmount assertions in cents (2100).

+++ server/internal/db/migrations/001_initial.sql
@@ -77,10 +77,10 @@ CREATE TABLE invoices (
     buyer_individual BOOLEAN,
     buyer_banks TEXT,
 
-    -- Totals
-    amount_without_vat REAL,
-    vat_amount REAL,
-    amount_with_vat REAL,
+    -- Totals (stored as integer cents to avoid float rounding drift)
+    amount_without_vat BIGINT,
+    vat_amount BIGINT,
+    amount_with_vat BIGINT,
     duplicate_of_id TEXT REFERENCES invoices(id),
     error_message TEXT,
 
@@ -93,9 +93,9 @@ CREATE TABLE invoice_items (
     invoice_id TEXT NOT NULL REFERENCES invoices(id) ON DELETE CASCADE,
     description TEXT,
     quantity REAL,
-    unit_price REAL,
-    total_price REAL,
-    vat_amount REAL,
+    unit_price BIGINT,
+    total_price BIGINT,
+    vat_amount BIGINT,
     vat_rate REAL,
     vat_classifier TEXT,
     created_at INTEGER NOT NULL DEFAULT (unixepoch())
